requirements.txt
asgiref==3.6.0
django-environ==0.9.0
Django==4.1.4
pip==21.3.1
psycopg2==2.9.5
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.7
wheel==0.37.1

django-crispy-forms~=1.14.0
crispy-bootstrap5==0.7